requirements.txt
python-dotenv
supabase
supabase
psycopg2-binary
yfinance
pandas
numpy
pandas-ta
requests
pytrends
textblob
snscrape
joblib
prophet
scikit-learn
ntfy
tqdm
fredapi
